* uim/uim-scm.h   - (uim_scm_symbol_value_int): Change return type from int to long * uim/uim-scm-sigscheme.c   - (uim_scm_symbol_value_int): Change return type from int to long   - (uim_scm_symbol_value_int_internal): Follow the change * doc/COMPATIBILITY   - Update "uim-scm API reorganization in uim 1.5.0"

+++ uim/uim-scm.h
@@ -118,7 +118,7 @@ void uim_scm_error_obj(const char *msg, uim_lisp errobj);
 /* evaluations */
 uim_lisp uim_scm_symbol_value     (const char *symbol_str);
 uim_bool uim_scm_symbol_value_bool(const char *symbol_str);
-int      uim_scm_symbol_value_int (const char *symbol_str);
+long     uim_scm_symbol_value_int (const char *symbol_str);
 char    *uim_scm_symbol_value_str (const char *symbol_str);
 
 uim_lisp uim_scm_eval(uim_lisp obj);
